Refactor `io` into cpu module
